pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowlyr-crew"
version = "0.1.43"
description = "AI Skill Loader - define professional skills in Markdown, load into Claude Code and other AI IDEs via MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Liu Kai", email = "mrliukai@gmail.com" }
]
keywords = [
    "ai-skill-loader", "mcp", "model-context-protocol",
    "digital-employee", "claude-code", "cursor", "ai-ide",
    "prompt-engineering", "ai-automation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
id = ["httpx>=0.25"]
avatar = ["Pillow>=10.0"]
execute = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.5"]
webhook = ["starlette>=0.27", "uvicorn>=0.20", "croniter>=2.0"]
trajectory = [
    "knowlyr-core>=0.1.0",
    "knowlyr-sandbox>=0.1.0",
    "knowlyr-recorder>=0.1.0",
    "knowlyr-reward>=0.1.0",
    "knowlyr-hub>=0.1.0",
]
dev = ["pytest", "pytest-asyncio", "ruff"]
all = ["knowlyr-crew[mcp,id,execute,openai,dev,avatar,webhook,trajectory]"]

[project.scripts]
knowlyr-crew = "crew.cli:main"

[project.entry-points."mcp.servers"]
knowlyr-crew = "crew.mcp_server:serve"

[project.urls]
Homepage = "https://github.com/liuxiaotong/knowlyr-crew"
Repository = "https://github.com/liuxiaotong/knowlyr-crew"
Issues = "https://github.com/liuxiaotong/knowlyr-crew/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/crew"]

[tool.hatch.build]
artifacts = [
    "src/crew/employees/*.md",
    "src/crew/employees/pipelines/*.yaml",
    "src/crew/employees/discussions/*.yaml",
    "src/crew/builtin_templates/*.md",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = [
    "E501",   # line too long
    "E402",   # module level import not at top (有意延迟 import)
    "E741",   # ambiguous variable name
    "B904",   # raise without from
    "B007",   # unused loop variable
    "B011",   # assert False
    "B017",   # assertRaises(Exception)
    "F841",   # local variable assigned but never used (TODO: 分批清理)
    "F821",   # undefined name (动态 import / TYPE_CHECKING)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
